# docs/swagger.yaml
basePath: /api
definitions:
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  model.MaterialWithCalculationData:
    properties:
      coefficient:
        type: number
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      is_deleted:
        type: boolean
      slope_angle:
        type: integer
      title:
        type: string
      volume_result:
        type: number
    type: object
  model.PitsCalculationListItem:
    properties:
      calculated_materials_count:
        type: integer
      completed_at:
        type: string
      created_at:
        type: string
      creator_id:
        type: integer
      formed_at:
        type: string
      id:
        type: integer
      moderator_id:
        type: integer
      pit_depth:
        type: number
      pit_length:
        type: number
      pit_width:
        type: number
      status:
        type: string
    type: object
  model.PitsCalculationWithMaterials:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      creator_id:
        type: integer
      formed_at:
        type: string
      id:
        type: integer
      materials:
        items:
          $ref: '#/definitions/model.MaterialWithCalculationData'
        type: array
      moderator_id:
        type: integer
      pit_depth:
        type: number
      pit_length:
        type: number
      pit_width:
        type: number
      status:
        type: string
    type: object
...
  model.UpdatePitParam:
    properties:
      pit_depth:
        type: number
      pit_length:
        type: number
      pit_width:
        type: number
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для расчета земляных работ при разработке котлована
  title: DIA API
  version: "1.0"
paths:
...
  /pits:
    get:
      consumes:
      - application/json
      description: Получение списка всех заявок с фильтрацией (только для модераторов)
      parameters:
      - description: Фильтр по статусу
        in: query
        name: status
        type: string
      - description: 'Начальная дата (формат: YYYY-MM-DD)'
        in: query
        name: start_date
        type: string
      - description: 'Конечная дата (формат: YYYY-MM-DD)'
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.PitsCalculationListItem'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Получение списка заявок
      tags:
      - pits
  /pits/{id}:
    delete:
      consumes:
      - application/json
      description: Удаление заявки (только для модераторов)
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Удаление заявки
      tags:
      - pits
    get:
      consumes:
      - application/json
      description: Получение детальной информации о заявке включая материалы
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PitsCalculationWithMaterials'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Получение заявки с материалами
      tags:
      - pits
    put:
      consumes:
      - application/json
      description: Обновление данных заявки (только владелец заявки)
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.UpdatePitParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Обновление заявки
      tags:
      - pits
  /pits/{id}/complete:
    put:
      consumes:
      - application/json
      description: Перевод заявки в статус "completed" или "rejected" (только для
        модераторов)
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: 'Статус: ''completed'' или ''rejected'''
        in: body
        name: input
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Завершение/отклонение заявки
      tags:
      - pits
  /pits/{id}/form:
    put:
      consumes:
      - application/json
      description: Перевод заявки из черновика в статус "formed" (только владелец
        заявки)
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Формирование заявки
      tags:
      - pits
  /pits/draft:
    get:
      consumes:
      - application/json
      description: Получение ID черновика заявки пользователя и количества материалов
        в ней. Для гостей возвращает -1.
      produces:
      - application/json
      responses:
        "200":
          description: 'Для авторизованных: {pit_id, pits_count}, для гостей: -1'
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Получение черновика заявки
      tags:
      - pits
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Введите JWT токен в формате: "Bearer {your_token}".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"